Cleanup some code only used in Go implementation
